+++ CHANGELOG.md
@@ -37,7 +37,8 @@
 - Enh #190: Use `str_contains` for case-sensitive match in `LikeHandler` (@samdark)
 - Enh #194: Improve psalm annotations in `LimitableDataInterface` (@vjik)
 - Bug #195: Fix invalid count in `IterableDataReader` when limit or/and offset used (@vjik)
-- Enh #201: Disable sorting when limit is set explicitly in a paginator (@samdark)
+- Enh #201: Disable sorting when limit is set explicitly in a paginator, use `InvalidArgumentException` when sorting
+  is supplied in this case (@samdark)
 
 ## 1.0.1 January 25, 2023